+++ test/models/test_user_model.dart
@@ -40,3 +40,5 @@ class TestUserModel {
     return "TestUserModel:$name";
   }
 }
+
+class UnserializableModel {}